fir_ser/requirements.txt
celery==5.1.0
django-celery-results==2.0.1
django-celery-beat==2.2.0
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-xml==2.0.0
django-simple-captcha==0.5.14
mysqlclient==2.0.3
django-redis==4.12.1
dnspython==2.1.0
oss2==2.14.0
aliyun-python-sdk-sts==3.0.2
qiniu==7.4.1
xmltodict==0.12.0
pyOpenSSL==20.0.1
paramiko==2.7.2
PyJWT==2.1.0
drf-extensions==0.7.0
python-daemon==2.2.3
psutil==5.6.6
flower==1.0.0
django-proxy==1.2.1
uWSGI==2.0.20
django-cors-headers==3.10.1
django-filter==21.1